Allow Intel, rename tmpi.lock for user and exit on mpi error
This should make the whole thing work with Intel MPI #11 as it no also
checks for `Intel` in the `mpirun --version` output. It also gets rid of
the `-pmi-port` which does not seem to be required for mpich anymore.
Any help on the usage is appreciated.

The tmpi.lock has been renamed to tmpi_$USER.lock so that there are no
issues on any multiuser systems.

When experiencing an issue with MPI startup, then the dummy window is
killed forcefully as it would just show an empty window and the actual
output is on the launch window. Related issue: #5